Refactor capnweb capabilities.
